move graph code to seperate files
